Deviceadapters: removed compiler warnings from multiple adapterors
git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@17200 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd

+++ DeviceAdapters/IsmatecMCP/MCPCommands.h
@@ -60,9 +60,9 @@ class MCPCommand
 protected:
    static int ParseDecimal(const std::string& s, int maxNDigits, int& result)
    {
-      if (s.empty() || s.size() > maxNDigits)
+      if (s.empty() || s.size() > (unsigned int) maxNDigits)
          return ERR_UNEXPECTED_RESPONSE;
-      for (int i = 0; i < s.size(); ++i)
+      for (unsigned int i = 0; i < s.size(); ++i)
       {
          char c = s[i];
          if (c < '0' || c > '9')
